When resampling off the image, return value is now configurable.

+++ contrib/mul/vil3d/vil3d_trilin_interp.h
@@ -55,12 +55,12 @@ inline double vil3d_trilin_interp_safe(double x, double y, double z, const T* da
                                        vcl_ptrdiff_t xstep, vcl_ptrdiff_t ystep, vcl_ptrdiff_t zstep,
                                        T outval=0)
 {
-  if (x<0) return outval;
-  if (y<0) return outval;
-  if (z<0) return outval;
-  if (x>=nx-1) return outval;
-  if (y>=ny-1) return outval;
-  if (z>=nz-1) return outval;
+  if (x<0) return static_cast<double>(outval);
+  if (y<0) return static_cast<double>(outval);
+  if (z<0) return static_cast<double>(outval);
+  if (x>=nx-1) return static_cast<double>(outval);
+  if (y>=ny-1) return static_cast<double>(outval);
+  if (z>=nz-1) return static_cast<double>(outval);
   return vil3d_trilin_interp_raw(x,y,z,data,xstep,ystep,zstep);
 }
